requirements.txt
google-cloud-documentai==2.27.0
python-dotenv==1.0.1
langchain==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langchain-openai==0.1.10